Simplify Popover documentation, stub Tooltips
